Fix POWOperation nonce type to support large values beyond int64 range
- Change POWOperation.Nonce from *Int to *UInt64 to match Steem protocol
  (nonce is uint64_t in C++ code)
- Add test cases to verify large nonce values can be parsed correctly
- Fixes unmarshal error for block 64500 where nonce value
  "11115213149964598312" exceeds int64 maximum

This resolves the issue where parsing operations in block 64500 failed with:
"strconv.ParseInt: parsing '11115213149964598312': value out of range"

+++ protocol/operations.go
@@ -212,7 +212,7 @@ func formatAssetFromObject(amount, nai string, precision uint8) string {
 type POWOperation struct {
 	WorkerAccount string           `json:"worker_account"`
 	BlockID       string           `json:"block_id"`
-	Nonce         *Int             `json:"nonce"`
+	Nonce         *UInt64          `json:"nonce"`
 	Work          *POW             `json:"work"`
 	Props         *ChainProperties `json:"props"`
 }
